requirements.txt
opencv-python
numpy
scipy
matplotlib
scikit-learn
Pillow
